api/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-dotenv==1.0.0
supabase==1.0.3
firebase-admin==6.2.0
requests==2.31.0
python-multipart==0.0.6
boto3==1.28.45 